refactor: add SecretRecoveryPhrase component, tweak copy button (#24)
